[API] (PC-12493) webapp: move serializer to native
